cast bool to int

+++ tests/regex.cs
@@ -7,6 +7,6 @@ public partial class RegExCompiled
   [SqlFunction(IsDeterministic = true, IsPrecise = true)]
   public static bool RegExCompiledMatch(string pattern, string matchString)
   {
-    return Regex.Match(matchString.TrimEnd(null), pattern.TrimEnd(null), RegexOptions.Compiled).Success;
+    return (int)Regex.Match(matchString.TrimEnd(null), pattern.TrimEnd(null), RegexOptions.Compiled).Success;
   }
 };
